requirements.txt
Django==1.9.4
djangorestframework==3.3.3
googlemaps==2.4.4
httplib2==0.9.2
oauth2==1.9.0.post1
requests==2.10.0
six==1.10.0
yelp==1.0.2
